Make objects sync
