+++ kr8s/asyncio/objects.py
@@ -1,0 +1,37 @@
+from kr8s._objects import (  # noqa
+    APIObject,
+    Binding,
+    ClusterRole,
+    ClusterRoleBinding,
+    ComponentStatus,
+    ConfigMap,
+    ControllerRevision,
+    CronJob,
+    CustomResourceDefinition,
+    DaemonSet,
+    Deployment,
+    Endpoints,
+    Event,
+    HorizontalPodAutoscaler,
+    Ingress,
+    IngressClass,
+    Job,
+    LimitRange,
+    Namespace,
+    NetworkPolicy,
+    Node,
+    PersistentVolume,
+    PersistentVolumeClaim,
+    Pod,
+    PodDisruptionBudget,
+    PodTemplate,
+    ReplicaSet,
+    ReplicationController,
+    ResourceQuota,
+    Role,
+    RoleBinding,
+    Secret,
+    Service,
+    ServiceAccount,
+    StatefulSet,
+)
