+++ kr8s/asyncio/portforward.py
@@ -1,0 +1,1 @@
+from kr8s._portforward import PortForward  # noqa
